Nearly complete Aerospace Blockset removal from Forces and Moments
Nearly complete removal of Aerospace Blockset from the Forces and
Moments block. I left both the old and the new in, hooked up with scopes
for comparions. I think I foung one error in the way the Aerospace
Blockset was implemented:

The Force coefficients are in wind, and the Moment coefficients are in
body. In the old Aerospace Blockset, the wind axes forces are crossed
with the moment arm to produce the moment contribution in the wind axis.
However, I think this is wrong, it should be body axis for the moment.
So, in the revised block, I cross body axis forces with the moment arm
to prouduce body axis moments. Please have a look and see if you agree.

If everyone agrees the new method is better, I will remove the dignosis
hoookup.

The DCM block used in this area is still Aerospace Blockset. I haven't
gotten around to creating a custom one yet. Once that is complete, the
forces and moments block will be completely open source.

I also re-saves the UAV_NL model with the new file extension type.