Adding Validated Dynamics Text File for UltraStick25e
Here is one way to keep track of the sim. I am adding the complete
printout from running the setup for the NL sim as a text file in the
aircraft confiuguration folder.

This might not be very elegant, and I'm sure we can automate this, but
hey, if you cause a change you didn't intend to make to the dynamics,
you would have a record of that when comparing to this file, and it
might prevent a runaway.